Fix mending bug
It was caused by an issue with ExtraVerificationCheck function, where it wrongly assumed that all repair ingredients do not use hit points, and that all ingredients that do use hit points are actually equipments to be mended. That's clearly not the case, as things like cloth do use hit points, and when those ingredients were being evaluated to be used for mending, the bill filtered them out as if they were equipments that were being mended.