Ensure prisoners are always recruitable (#2)
Alfred applies a flat 0.2 multiplier to the calculated resistance
reduction on each warden->prisoner interaction, but if the original
reduction value exceeds the remaining resistance of the prisoner, then
Interacted() will have already limited it to not exceed the remaining
resistance. The effect is that the prisoner becomes unrecruitable,
as each interaction will only reduce their remaining resistance by 20%,
never letting it reach zero.

As a fix, apply the multiplier to the original resistance reduction,
before it's checked and limited against the remaining resistance.